+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Unreleased
 
 - **[NEW]** Show month corresponding to `initialDate` in `DatePicker`
+- **[FIX]** Fix random font-size issue in `Stepper` using `StepperDisplay.LARGE`
+- **[NEW]** Add `id` prop to `Title`
+- **[NEW]** Add `id` prop to `TheVoice`
+- **[NEW]** Add focus styles to `ItemRadio` (a11y)
+- **[NEW]** Add `ariaLabelledBy` prop to `ItemRadioGroup` (a11y)
 - [...]
 
 # v15.0.4 (20/11/2019)